+++ NEWS.md
@@ -1,6 +1,8 @@
 # dplyr 0.4.3.9000
 
-* `group_by` supports `column` (#1012). 
+* `bind_rows` handles 0 length named list (#1515). 
+
+* `group_by` supports `column` (#1012).
 
 * `mutate` on a grouped data does not drop grouping attributes (#1120).